feat : FMP 에서 주식 목록, 시가총액 가져오기

+++ src/main/java/redlightBack/stock/dto/FavoriteStockResponse.java
@@ -1,8 +1,10 @@
 package redlightBack.stock.dto;
 
+import java.math.BigDecimal;
+
 public record FavoriteStockResponse(
         String name,
         String code,
-        Long marketCap
+        BigDecimal marketCap
 ) {
 }
